Add unit tests for tachymtry - part 2. Some small fixes

+++ src/main/java/points/transforming/app/server/models/tachymetry/api/TachymetryMetaDataResponse.java
@@ -9,5 +9,5 @@ public class TachymetryMetaDataResponse {
     private final String tachymetrName;
     private final String tachymetrType;
     private final Double temperature;
-    private final Integer pressure;
+    private final Long pressure;
 }
